Fixed the event queue when running linux or real system that dont forget to implement sem_init (Yeah, you OSX).

+++ src/lib/cli_cpp/cmd/A_Command.cpp
@@ -95,7 +95,7 @@ bool A_Command::Is_Argument_Substring( const int&         argument_index,
                                        std::string&  match_name )const
 {
     // Don't go over bounds
-    if( argument_index >= m_command_argument_list.size() ){
+    if( argument_index >= (int)m_command_argument_list.size() ){
         return false;
     }
